Support browser via wasm target
