+++ gradle.properties
@@ -1,4 +1,17 @@
+#Gradle
+org.gradle.jvmargs=-Xmx2048M -Dfile.encoding=UTF-8 -Dkotlin.daemon.jvm.options\="-Xmx2048M"
+org.gradle.caching=true
+org.gradle.configuration-cache=true
+
+#Kotlin
 kotlin.code.style=official
-kotlin.version=1.9.21
-agp.version=8.2.0
-compose.version=1.5.11
+kotlin.js.compiler=ir
+
+#Android
+android.useAndroidX=true
+android.nonTransitiveRClass=true
+
+#Compose
+org.jetbrains.compose.experimental.uikit.enabled=true
+org.jetbrains.compose.experimental.jscanvas.enabled=true
+org.jetbrains.compose.experimental.wasm.enabled=true
