+++ src/commonMain/kotlin/game/GameOfLifeFactory.kt
