+++ src/commonMain/kotlin/game/util/RandomNumberGenerator.kt
@@ -1,10 +1,12 @@
 package game.util
 
+import kotlin.random.Random
+
 interface RandomNumberGenerator {
     fun nextDouble(): Double
 }
 
 class RandomNumberGeneratorImpl : RandomNumberGenerator {
 
-    override fun nextDouble() = Math.random()
+    override fun nextDouble() = Random.nextDouble()
 }
