+++ src/commonMain/kotlin/ui/model/ClickPosition.kt
@@ -1,6 +1,7 @@
 package ui.model
 
 import androidx.compose.ui.geometry.Offset
+import kotlin.jvm.JvmInline
 
 @JvmInline
 value class ClickPosition(val offset: Offset)
