+++ src/commonMain/kotlin/ui/model/PivotPointPosition.kt
@@ -1,6 +1,7 @@
 package ui.model
 
 import androidx.compose.ui.geometry.Offset
+import kotlin.jvm.JvmInline
 
 @JvmInline
 value class PivotPointPosition(val offset: Offset)
