+++ src/wasmJsMain/kotlin/main.kt
@@ -1,17 +1,19 @@
+
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.window.*
-import ui.composable.Application
+import androidx.compose.ui.ExperimentalComposeUiApi
+import androidx.compose.ui.window.CanvasBasedWindow
 import game.GameOfLifeFactoryImpl
 import game.model.GameStateFactoryImpl
 import game.model.GridFactoryImpl
 import game.util.RandomNumberGeneratorImpl
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
+import ui.composable.Application
 import ui.model.GameUiStateFactoryImpl
 import ui.viewmodel.GameOfLifeViewModel
 
+@OptIn(ExperimentalComposeUiApi::class)
 fun main() {
     val gameScope = CoroutineScope(Dispatchers.Default)
     val viewModelScope = CoroutineScope(Dispatchers.Default)
@@ -22,15 +24,8 @@ fun main() {
     val gameUiStateFactory = GameUiStateFactoryImpl(gameStateFactory)
     val viewModel = GameOfLifeViewModel(gameFactory, viewModelScope, gameUiStateFactory)
 
-    application {
+    CanvasBasedWindow("YAIoGol") {
         val gameUiState by viewModel.gameUiStateFlow.collectAsState()
-
-        Window(
-            onCloseRequest = ::exitApplication,
-            title = "Generation ${gameUiState.gameState.generation}",
-            state = rememberWindowState(width = 1200.dp, height = 800.dp)
-        ) {
-            Application(viewModel, gameUiState)
-        }
+        Application(viewModel, gameUiState)
     }
 }
