+++ src/wasmJsMain/resources/index.html
@@ -1,0 +1,14 @@
+<!doctype html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>YAIoGol</title>
+    <script src="skiko.js"></script>
+</head>
+<body>
+<div>
+    <canvas id="ComposeTarget" width="800" height="600"></canvas>
+</div>
+<script src="YAIoGoL.js"></script>
+</body>
+</html>
